Changed ByteArray's pending_word_len to usize (#3789)

+++ corelib/src/test/byte_array_test.cairo
@@ -247,7 +247,7 @@ fn test_concat_pending_sum_up_to_more_than_word_gt16() {
 
 use debug::PrintTrait;
 fn compare_byte_array(
-    mut ba: @ByteArray, mut data: Span<felt252>, pending_word_len: u8, pending_word: felt252
+    mut ba: @ByteArray, mut data: Span<felt252>, pending_word_len: usize, pending_word: felt252
 ) {
     assert(ba.data.len() == data.len(), 'wrong data len');
     let mut ba_data = ba.data.span();
